Size-specific free list in HashStringAllocator (#5487)
Summary:
We maintain size specific free lists for faster allocation. When splitting or merging blocks we may move from one free list to another. For allocations that do not have to be contiguous, we check the size specific or larger lists first, then consume from the smaller ones before growing the HashStringAllocator.

For allocations in the MmapAllocator size classes size range, we go directly to the allocator of the pool and do not use the HashStringAllocator arenas. In this way, when such a block is freed it will become available to the rest of the system, which would not happen if this were embedded in an extra large HashStringAllocator arena.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5487

Reviewed By: mbasmanova

Differential Revision: D47160562

Pulled By: oerling

fbshipit-source-id: b4e51a4c5e061a788b402d69222284770de569af