Add RowContainer::extractNulls API (#5502)
Summary:
extractNulls API retrieves all positions with null values in a vector for the RowContainer.

This API is used for ignoreNulls logic in Window functions. Columns with null values are skipped during function evaluation in those cases. This API is a convenient way to determine the null positions without column value extraction.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5502

Reviewed By: Yuhta

Differential Revision: D47239730

Pulled By: mbasmanova

fbshipit-source-id: af804464ef7cf18c4a1122540d28d624d64bcae9