Fix fileOffset in ParquetRowReader::filterRowGroups() (#5507)
Summary:
ParquetRowReader::filterRowGroups() get the offset of the first data or dictionary page to the beginning of the file and assign it to fileOffset, which would be used to see if this RowGroup is in the split's range. It uses the ColumnChunk.file_offset (rowGroups_[i[.columns[0].file_offset) as the fileOffset when the Rowgroup's file_offset
(rowGroups_[i].file_offset) was not set, but ColumnChunk.file_offset is actually the offset to the ColumnMetaData, not the offset to the data or dictionary pages. Furthermore ColumnChunk.file_offset is optional, which would cause queries to fail. This commit fixed the fileOffset calculation to use ColumnMetaData.data_page_offset or ColumnMetaData.dictionary_page_offset, whichever is first to occur. The ColumnMetaData.data_page_offset is guaranteed to be existing.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5507

Reviewed By: mbasmanova

Differential Revision: D47293406

Pulled By: Yuhta

fbshipit-source-id: 392cbc13da0528f3c6e5f4c57b7cb7633fd6849f