Add support in HiveParser to identify Date logical type (#5562)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5562

Date was recently transitions to being a logical type instead of a TypeKind. This meant that HiveParser would translate 'date' hive type to an INTEGER velox type which corresponds to DATE's physical type. However this violated a check in TableHandler which compared velox and hive types. This change ensures that 'date' hive type is correctly converted to it DATE velox type.

Reviewed By: xiaoxmeng, amitkdutta

Differential Revision: D47301882

fbshipit-source-id: aee68a637a80aac828fd4f5a679d4d903fd7b924