Extend hive partition function and spec to support multi table writer drivers (#5557)
Summary:
For bucketed table write, we need a local partition between remote exchange
and table writers. The local partition needs to repartition data based on the
bucket id to ensure the data with the same bucket id go to the same table writer.
The partition function should use the hive partition function but the bucket to
partition map is unknown when we convert the query plan (from presto to velox).
So can only create hive partition spec with the number of buckets. When task
instantiates the drivers, the local planner will pass the actual number of table
writer drivers as the number of partitions to create the hive partition function used
by the local partition operator. The hive partition function can use round robin to
map from bucket to partition as extended by this PR.
This PR also changes the hive data sink to use hive partition function instead of
hash partition function be consistent with hive partition function used by local
partition operator. Since hive data sink doesn't need any mapping on the bucket
id, so we extend a hive function to handle the case without the bucket to partition
map. But the alternative is to let the hive data sink to build an identical map and
pass to the hive partition function.

Details see [multi driver support](https://github.com/facebookincubator/velox/issues/5546)

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5557

Reviewed By: mbasmanova

Differential Revision: D47287813

Pulled By: xiaoxmeng

fbshipit-source-id: 155b958cb75499be660254eee97a2f104c9618bd