Allow loading of nested lazy RowVector (#5392)
Summary:
Followup of: https://github.com/facebookincubator/velox/pull/5140

For cases when we have nested lazies in row vector, we should load them appropriately when decoding, or calling loadedvector(). This PR addresses this loading for lazies in row type.

I'm not sure if its safe to mutate the RowVector when loading.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5392

Reviewed By: bikramSingh91

Differential Revision: D47172716

Pulled By: pranjalssh

fbshipit-source-id: c27c3bf5e99a510d88459e276dca03b697bc59da