Add hash table benchmark (#5582)
Summary:
the benchmark makes different sizes of hash join tables and probes these with different hit rates. In cases that end up with kNormalizedKey tables, this also repeats the probe with F14FastSet to provide a comparison.

We separately measure the time to hash and the time to probe.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5582

Reviewed By: Yuhta

Differential Revision: D47341674

Pulled By: oerling

fbshipit-source-id: f9838bff0cffa83e88fed6cbe20920430f83b616