[native pos] Propagate taskInfo updates to Spark metrics
Currently, we update the spark metrics accumulators
only after the task is complete. This means that spark
sees no progress or activity on the task until the task
is complete. This also means that we can't effectively use
the stalled tasks killer in spark scheduler as it relies on
spark accumulators being updated regularly to measure if a
task is making progress.

This commit fixes this by ensuring that everytime we receive
TaskInfo from CPP process we immediately update Spark accumulators
to let spark know that we are making progress