Cancle Benchmark job on new commit (#5636)
Summary:
Saves CI resources on repeated commits e.g. due to linting.

The group key influences which jobs will be canceled. By adding the workflow name we ensure that only other benchmark jobs will be canceled. `${{ github.head_ref || github.sha }}` Is necessary as this is run on `pull_request` and `push` as `head_ref` is the source branch for a PR but empty on `push` so we will not cancel consecutive runs on push to main as the commit sha is part of the key there instead of the branch name.

Closes https://github.com/facebookincubator/velox/issues/5623

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5636

Reviewed By: Yuhta

Differential Revision: D47436166

Pulled By: kgpai

fbshipit-source-id: 10bb142c6947fc72e094afad12ae0e54f1a9ce0f