[native pos] Fix warning on failure to update metrics
I noticed a warning in the logs:

PrestoSparkStatsCollectionUtils An error occurred while updating Spark Internal
metrics for task=TaskInfo
{taskId=20230713_154415_00000_es9u8.0.0.0.0, state=FINISHED}

There was not error message though. Turns out, the exception was passed
incorrectly to log.warn(). Fixed.

The cause the error was that TaskContext is null. Added a check to avoid the
warning.