[native pos] Fix TaskInfo fetcher logic to avoid concurrent runs
Currently, due to using async callback inside the already async scheduled thread
we are making the scheduleAtFixedDelay's delay mechanism ineffective because
we early exit from the doServerGet call. The end result is that we will have
multiple outstanding GET/info call at a given time.

Making this call sync on the thread allows us to guard against this by ensuring
that we do not start the next instance of the work (http call) before
the first one returns.