Implement iceberg.use_sample_statistics config (#4)
* Implement iceberg.use_sample_statistics config

When enabled, this config will return table statistics calculated
from the $sample meta table, if it exists.

Additionally, this includes support for using the latest version of
the sample table corresponding to the queried snapshot table

An example of how time travel works for samples is detailed by
the following example:

Timeline:

orders:   |--t1----------t3------t5-----t6---------------->
$sample   |----t2----------t4----------------------------->

In this case, if call getTableStatistics at for "orders" with samples
enabled at time:

- t1: uses $sample @ t2
- t3: uses $sample @ t2
- t5: uses $sample @ t4
- t6: uses $sample @ t4

Put into words, the version of the $sample table used will always
correspond to the oldest sample before the time at which the
snapshot for the queried version of the table was taken.

This has a drawback of tables never actually using the most
recent version of the $sample table unless the snapshot queried
is after the table. This behavior could change in the future.

* Remove use of SinglePathCatalog

replaced with use of inner class that only functions to allow the
catalog to be used in try-with-resources statements.

* Rename create sample table procedure

* Add tests for create, delete, insert, getStats

- create new DeleteSampleTableProcedure

+basic tests to ensure:

- queries which create/delete samples can execute properly
- queries which insert into the sample table execute properly
- queries which query the sample table execute properly
- queries which call getStatistics and use the proper config will
pull statistics from the correct table

* Drop sample table when actual is dropped

* Changelog delta calculation to pick sample snap

See the javadoc inside TableStatisticsMaker for a description of
the algorithm