Fix unmap sizes (#5662)
Summary:
When a mmap is made with a different size and maxsize, the unmap must unmap maxsize. But the utilization must be decremented by size. Failure to do this explodes the number of mmaps and fragments the address space by leaving unused tails of large mmaps hanging after unmapping the head.

To validate the effect, the test must access the process rss. Does this with Linux specific ps.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5662

Reviewed By: xiaoxmeng

Differential Revision: D47502982

Pulled By: oerling

fbshipit-source-id: b1e7c053fffc3b0271bb7b1869bf62faa1390ff1