Fix final selection handling while executing lamda expressions (#5628)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5628

Currently, the final selection passed down to lambdas is used as a proxy to preserve rows in a reused result vector that are not being operated on in the current callable. This causes confusion for authors when they implement vector functions that use lambdas and they instead end up passing the actual final selection in the context. When this happens, the evaluation loop inside the lambda can run into errors when it executes code that uses final selection. This is because lambdas can take in both results from evaluated expressions and top level fields as input where the final selection rows may not apply to the former as they can have different sizes or valid rows
This patch attempts to fix this by explicitly specifying and using the selectivity vector for preserving unused but valid rows. It also corrects its usage across all vector functions that use lambdas as inputs.

Reviewed By: mbasmanova

Differential Revision: D47417956

fbshipit-source-id: 8c80651f61cb05c7f78cfd5619054d507191ed89