Fix leak of HashStringAllocator external alloc (#5689)
Summary:
Destruction/clear() of HashStringAllocator must free allocations it made directly from its pool(). Normally we do not see this because uses with aggregates destruct the accumulators before freeing the HSA. In this way the external allocations are freed when freeing the accumulator and are not left hanging until the HSA itself is cleared. But HSA semantics still require free of unfreed external allocations on destruction since whether an allocation is external or not is transparent to the user.

Test verifies that clear() frees external allocations. Prior to the change, only destruction did free these.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5689

Reviewed By: bikramSingh91

Differential Revision: D47533478

Pulled By: oerling

fbshipit-source-id: d36f372c3e8668bedc97fe0071bd24ac9706f1a1