Implement Aggregate::toIntermediate for map_agg (#5671)
Summary:
Implement Aggregate::toIntermediate for map_agg to optimize
abandon-partial-aggregation use case.

This is the first implementation of the API. Not surprising, it exposed some
bugs in the aggregation framework.

GroupingSet::abandonPartialAggregation clears the hash table causing destruction
of the HashStringAllocator that was provided to the aggregate function. This
causes crashes when Aggregate::toIntermediate tries to access the allocator.
Fixed that.

Also, the API documentation wasn't clear about how to interpret the 'rows'
parameter. Updated the documentation.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5671

Reviewed By: xiaoxmeng

Differential Revision: D47517563

Pulled By: mbasmanova

fbshipit-source-id: c5374641d578c4dc7a12e62a7d26ccf23d02ca5a