Disable Vsize/RSS check (#5712)
Summary:
Vsize and RSS reflect actions of the program but ps may not report these either synchronously or the OS may round up the quantities. So asserting changes of Vsize or RSS in a test may depend on OS version, load situation, retrograde Mercury and other factors not controlled by unit test. The concept of checking that mmap and munmap match is valid but experience shows this cannot run predictably in an unknown test environment.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5712

Reviewed By: xiaoxmeng

Differential Revision: D47579678

Pulled By: laithsakka

fbshipit-source-id: 08161c3ef848d34008220305fa628e0ee0b0fde5