Add SIMD-based implementation of json_extract_scalar (#5527)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5527

Continuing with the work in https://github.com/facebookincubator/velox/pull/5019 and building on
https://github.com/facebookincubator/velox/pull/5525 this adds a version of the JsonExtractScalar UDF with
the underlying implementation swapped from folly to simdjson.

This continues to show a significant performance improvement.  It also addresses an inconsistency between
the Velox and Presto Java versions of the UDF where large numbers that would cause an integer overflow are
returned directly as strings.

Reviewed By: laithsakka

Differential Revision: D47240465

fbshipit-source-id: a55e179c59e3d93563967112767149490a32b831