Implement JsonExtractor with simdjson (#5525)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5525

This adds a new implementation of the JsonExtractor, called SIMDJsonExtractor, that provides similar
functionality, but using the more performant simdjson library in place of folly.

One key difference with simdjson that I had to work around is that it uses a cursor so values have a "use it or
lose it" ephemeral quality.  To handle this, I allow the user to pass in a consumer function that gets called on
final values (what would be returned by the original JsonExractor) while they're still valid.  I also had to make
the extract call recursive to handle the "*" operator on arrays.

I will update the json_extract and json_extract_scalar functions to use the new extractor in forthcoming diffs,
and include benchmark results there, but from what I've seen, the performance benefits are in line with what
we've seen from using simdjson in other JSON UDFs.

Ultimately, the plan is that this should entirely replace the JsonExtractor so we can remove it from the
codebase.

Reviewed By: laithsakka

Differential Revision: D47235020

fbshipit-source-id: 0a76aa89d1d752327d0a67be840a4cd4d20eb382