Fix flaky failure in TableScanTest.readMissingFieldsFilesVary (#5775)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5775

We've seen readMissingFieldsFilesVary fail in CI due to the null check at line 2762.

I have been unable to reproduce locally (I've run over 100k times), but my best guess is that the order we
process the splits isn't deterministic (e.g. due to concurrency at some level).  I wasn't sure if this would be
the case when I wrote the test.

I fixed the test to handle reading the splits out of order (the order isn't critical to the test), in the hopes that
this is the issue.

Reviewed By: laithsakka

Differential Revision: D47678569

fbshipit-source-id: 2c9de3de91314a9d9ec2b5dd4c47c1b2a6f6f079