Fix aggregate merge containers and chekc RowContainer variable length… (#5778)
Summary:
… free

In the case of merging after spilling, we have a source RowContainer and a merge RowContainer that share the same Aggregates. Therefore they must also share the same HashStringAllocator since this is referenced from the Aggregates. Adds a constructor parameter for sharing the stringAllocator of another RowContainer.

We add a flag to check that deleting the rows from a RowContainer causes its stringAllocator to go empty. This is needed in order to enforce that aggregates, when spilling, free the space they should in the container.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5778

Reviewed By: xiaoxmeng

Differential Revision: D47684831

Pulled By: oerling

fbshipit-source-id: b53a23c1f721c2cd0d39bc7880aa2a0ead64ce42