Implement Iceberg changelog scanning (#5)
* Implement Iceberg changelog scanning

This commit exposes a new system table called $changelog that
allows users to return a table pertaining to info about when
which snapshots specific records were added or deleted.

The schema of the changelog table is as follows:
| primary_key: T | operation: VARCHAR | ordinal: bigint | snapshotid: bigint |

To use this feature, the only requirement is that users need to specify
the name of a column in the table properties as the "primary key". The
value of the key will be exposes in the "primary_key" column of the
changelog table. The type of the column will be the same type as in
the underlying table.

The current beahvior in this commit only returns the results of the
changelog between the current and previous snapshot of the table being
queried. You can use this information to determine when a record has
been inserted or deleted.

SELECT committed_at, change.operation, primary_key, ordinal
FROM "orders$snapshots" as snap
JOIN "orders$changelog" as change
ON change.snapshotid = snap.snapshot_id;

               committed_at                 | operation | primary_key | ordinal
---------------------------------------------+-----------+-------------+---------
 2023-07-19 13:03:03.207 America/Los_Angeles | INSERT    |        1956 |       0
 2023-07-19 13:03:03.207 America/Los_Angeles | INSERT    |       28196 |       0
 2023-07-19 13:03:03.207 America/Los_Angeles | INSERT    |       27873 |       0

At a high-level the implementation strategy needed to expose information
from Iceberg's IncrementalChangelogScan tasks which differs from a
normal TableScan. The plan tasks for IncrementalScans include the
information on operation type, ordinal, and snapshotId which are are not
contained in the normal table scan plan.

When the connector metadata queries for a changelog table type, it
returns a schema with a static 4 columns, where the first 3 are static
VARCHAR, BIGINT, BIGINT, and the final is dependent upon the user-
defined primary key property.

To accomplish this the following changes were made
- a new primary key table property was added
- logic to generate the changelog table schema
- a new piece of info on iceberg splits containing the original table's
column handles which is used during split sourcing
- a new page source which constructs the correct format page based on
the extracted primary key.
- finally, some additional refactoring of the IcebergNativeMetadata and
IcebergHiveMetadata classes to consolidate certain bits of logic for the
ConnectorMetadata

And of course, this is preliminary work. There is still some more
optimization that could be done.

TODOs
- Add tests
- support table predicate
- support more than just parent ancestor changelog
- filter on a predicate when generating the incremental change log scan
- optimize to scan only for the primary key rather than the whole row

* Generate changelog page based on desired columns

Previously, the ChangelogPageSource always generated a static page with
the same 4 blocks inside. This behavior was incorrect. The page source
must only generate pages matching a given projection specification
which comes from the columns in the PageSourceProvider interface.

Before this change, only `SELECT *` queries worked. After this fix,
queries such as `count(*)`, and `SELECT primary_key` also work.

* Push predicate filters down to table scan

This commit fixes two issues

1. Predicate filters on the table scan columns such as 'ordinal', 'operation',
or 'snapshot_id' are not pushed into the incremental changelog scan as
they result in error since they aren't actual columns on the table
2. A predicate in the query corresponding to the primary key on the
changelog table are converted in the corresponding TupleDomain
referring to the IcebergTableHandle for the actual table. We now push
this predicate into the table scan to prevent scanning data during the
incremental scan which doesn't match the predicate

* Prevent scanning additional data for changelog

Since the changelog only surfaces the primary key, we don't need the
delegate page source to scan all of the other data. Previously the
implementation would scan for all of the rows since we always passed the
full set of column handles.

This change filters the set of column handles passed down to just the
primary key which should limit the amount of data scanned when querying
the changelog table.

* Add tests for changelog table

Found two issues while writing the tests and fixed them

- DeleteTableSampleProcedure now unsets the primary key and snapshot
properties when the sample table is deleted.
- TableStatisticsMaker throws an NPE in a particular case if there is no
snapshot ancestor available. We check for this case now to prevent the
NPE

* Fix sampling tests

Calls to create_table_sample need a new parameter since we started
requiring a primary key

* Fix tests for maven build

Maven runs tests in parallel which causes failures when
the tests use the same table name. Updated tests to
use names unique to this query runner.

* Add short circuit when getting system tables for changelog/samples

* throw TableNotFound when getting system table meta

This fixes the system table tests and also ensures that system table
meta is retrieved from the correct location

The code is interesting in that returning the system
table metadata is somewhat implicit. The regular connector metadata
can override the metadata from the returned system table's returned
schema.