Make row number generation optional in RowNumber operator (#5752)
Summary:
Changes in the Presto coordinator, made with PR
https://github.com/prestodb/presto/pull/19835 allows the row number generation to be skipped for a partial RowNumber node that feeds the Exchange operator. It reduces the amount of data sent.

The implementation is similar to the TopNRowNumber node and operator capability to skip row generation.

A subsequent PR is needed for Prestissimo to make use of the new capability when converting the Presto plan into the Velox plan.

This resolves the end-to-end test issue discussed in issue https://github.com/prestodb/presto/issues/20227.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5752

Reviewed By: xiaoxmeng

Differential Revision: D47727035

Pulled By: mbasmanova

fbshipit-source-id: c5f02dd7acbca1a91b395519a2d123f5d0318c60