Fix GroupingSet::toIntermediate fast path (#5825)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5825

When multiple aggregates exist and only some of them support
`toIntermediate`, we still need to construct the `RowContainer` for those who do
not support `toIntermediate`.  In this case when we clean up the `RowContainer`,
we try to destroy the rows using the aggregates with fast path by mistake and
result in bad memory access.  Fix this by creating a new `RowContainer` that
contains only aggregates that do not support `toIntermediate`.

Reviewed By: mbasmanova

Differential Revision: D47760457

fbshipit-source-id: 06050d1fae509da4716b23addaff2aa67397d2eb