Adding support for ALL_VALUES_MATCH
Returns whether all values of a map match the given predicate. Returns true if all the values match the predicate (a special case is when the map is empty); false if one or more values don’t match; NULL if the predicate function returns NULL for one or more values and true for all other values.