Store Parquet Logical Type information within ParquetTypeWithId (#5787)
Summary:
This PR integrates Parquet's LogicalType into ParquetTypeWithId. This LogicalType information will be used in choosing the correct Parquet Column Reader, determined by various factors contained in the LogicalType, such as the sign of the parquet column, the bit width, and the specific Logical Type of the column. This enhancement is among the necessary changes for issue https://github.com/facebookincubator/velox/issues/5770.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5787

Reviewed By: mbasmanova

Differential Revision: D47797593

Pulled By: Yuhta

fbshipit-source-id: 82bcda9a21e3e614d07a302f780d0bfd1d055e24