Fix flaky test
The test was not waiting for queries to be in running/queued state before moving forward.
Given the nature of DispatchManager it takes some time for queries to move to the right state.
This resulted in flaky behavior. As part of the fix, we are waiting for the queries to be in right
state before moving forward and that fixes the flakiness of the test.