Add OOM protection for table creation in hash build (#5780)
Summary:
The hash table itself can take a large amount of space when
prepare join table after all the hash build operators finish. This
step is non-reclaimable as we don't support spilling at that stage
if the memory allocation fails, the query might fail. Add a step to
reserve memory in that step to trigger spilling if the memory
reservation fails.

Also add utility to hast table to estimate the new hash table size
given the number distinct which is used by hash build to ensure
memory for table build.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5780

Reviewed By: oerling

Differential Revision: D47687086

Pulled By: xiaoxmeng

fbshipit-source-id: 99b4e0bdbd68cf654923a5ce4a4c0dfb7e5c3af5