Fix trackRowSize in SimpleAggregateAdapter (#5957)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5957

The RowSizeTracker used in SimpleAggregateAdapter used to be destroyed before the groups
are updated. This is because they are defined in if-constexpr blocks and if-constexpr are
converted to the if-body within a block during compilation, so the lifetime of RowSizeTracker is
limited. This diff fixes this problem by defining RowSizeTracker outside the if-constexpr blocks.

Reviewed By: Yuhta

Differential Revision: D47980733

fbshipit-source-id: f33ca882e9cf51f5616e57d6edee4b820f430eb9