[native] Change json location for reuse with other third parties
Nlohmmann's json.hpp is included in the source tree rather than a third party
 install.
The path used cannot be used in lieu of a system installation.
Thus, changing the location of the json.hpp to match the path found when
installed into the system.

This allows reuse of the already included library by other
third party includes that expect it to be installed.