Add function names to ColumnStatisticType
This change adds function name parameters to the enum values of
ColumnStatisticType. The function names are used when generating
ColumnStatisticMetadata for the ANALYZE query.

This change is in preparation to allow connectors to override the
function used to execute the statistics aggregations. Mainly it is to
support connectors that have differing underlying histogram or NDV
representations. For example, Hive natively generates KLL sketches
from Apache Datasketches for histograms, Spark has it's own custom
format. Some other connectors like mysql use a JSON format. Iceberg
tables can store NDV estimates in Apache DataSketches Theta sketches.
If we intend to support a variety of stats from other connectors this
change is necessary.

Allowing the connector to override the function for each statistic type
will let us generate statistic data in the format needed by the connector
in order to store it in the connector-specific catalog metadata.