dashboard: update webpack to v5
Node v16 is going to sunset on Sep 11 2023. Although
we only use it to compile the jsx files, it's still good to
switch to Node v18. Therefore, I also updated the
followings to newer versions:
- webpack and webpack-cli
- babel
- react

The extra benefit is:
The size of the new js files is reduced to almost 1/2.

While verifying the change, I also fixed some minor issues in
StageDetail.jsx and timeline.html.

Signed-off-by: Yihong Wang <yh.wang@ibm.com>