[native] Print velox user error log upon failure during start up.
If presto start up fails due to a user error (e.g. config file
is missing), the exception is not printed and server exits without
any trace of what happened. When this code was written, throwing
VeloxUserError was printing the error. However, later this
behavior is altered (https://github.com/facebookincubator/velox/pull/5610).
Hence the error needs to be logged for better debugging of incidents
where config files are missing.