Remove check for nulls in DictionaryVector::wrappedIndex() (#6394)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6394

A check is added recently requiring that DictionaryVector::wrappedIndex() is not called on indices where the dictionary vector is null. This check is too general since we have legit callsites of this function on null rows in the ContainerRowSerde class. This causes min() and max() aggregation functions to fail in the fuzzer test. This diff removes this check until we're able to identify and change all callers of wrappedIndex() to follow that expectation.

This diff fixes https://github.com/facebookincubator/velox/issues/6393.

Reviewed By: laithsakka

Differential Revision: D48914410

fbshipit-source-id: 1f3ba8f6858dd35958e7dfb6b623d33977869e10