Attribute lazy load times to table scan (#6460)
Summary:
When recording addInput, getOutput and finish times for Operators, subtracts the lazy load time accrued to the operator since last timing update. Adds the lazy load CPU and wall time to the source operator of the pipeline. Tracks lazy load with both CPU and wall time since this must be compatible with the CPU and wall time tracking in operator timing.

The runtime stats for lazy load are still kept at the level of the operator triggering the load. These times do not contribute to the operator's time as reported in addInputTiming and related.

Fixes https://github.com/facebookincubator/velox/issues/6357

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6460

Reviewed By: mbasmanova

Differential Revision: D49060047

Pulled By: oerling

fbshipit-source-id: d4b8739e8e69c98ea2ee111db4cf18d97e3e8295