Attribute lazy load times to table scan (#6558)
Summary:
When recording addInput, getOutput and finish times for Operators, subtract the lazy load time accrued to the operator since last timing update. Adds the lazy load CPU and wall time to the source operator of the pipeline. Tracks lazy load with both CPU and wall time since this must be compatible with the CPU and wall time tracking in operator timing.

The runtime stats for lazy load are still kept at the level of the operator triggering the load. These times do not contribute to the operator's time as reported in addInputTiming and related.

This is second time check-in of https://github.com/facebookincubator/velox/pull/6460

Fixes https://github.com/facebookincubator/velox/issues/6357

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6558

Reviewed By: oerling

Differential Revision: D49250981

Pulled By: gggrace14

fbshipit-source-id: d8fd100ec6bc9d6d6be8b34c482854ef4635c547