Update the offset when we filter out all values of a Row column in SelectiveStructColumnReader (#6575)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6575

SelectiveStructColumnReader has logic to exit early if none of the values pass a pushed down filter.
Unfortunately, we're missing the updates to the offsets when we do this.  This results in us either trying to re-
skip the values in the children when we read the next batch or reading to far ahed in the nulls buffer for the
Row which leads to attempting to read off the end of the buffer when we get to the end of the stripe.

The fix is just to make sure we update the offsets before returning.

Reviewed By: Yuhta

Differential Revision: D49285600

fbshipit-source-id: ed370b3702ad771c2885ab71f481fcfd1ee39084