Support SparkSQL rand function with seed specified (#6112)
Summary:
In Spark, rand() function can be called with a seed.

https://github.com/apache/spark/blob/7866da9519a586e617a55702c39f19d0e16b7279/sql/catalyst/src/main/scala/org/apahe/spark/sql/catalyst/expressions/randomExpressions.scala#L77

The query engine (Spark) is responsible for translating rand(seed) into rand
(seed, partitionIndex), where partitionIndex is a unique value assigned to each
thread of execution in a deterministic way. This ensures that different threads
produce different random values, but the query results are deterministic. The
query engine (Spark) also needs to ensure that data is partitioned across
threads of execution in a deterministic way.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6112

Reviewed By: Yuhta

Differential Revision: D49273878

Pulled By: mbasmanova

fbshipit-source-id: 86c2ff36362d5319f4f505e9e1e41a280b421ec2