Add driver thread context to fix memory arbitration deadlock #6625 (#6625)
Summary:
Meta internal shadowing test discover the following memory use case
can cause memory arbitration deadlock:
T1: opA allocate a buffer
T2: opB from a different driver re-allocate the buffer
T3: opB triggers the memory arbitration and suppose opB's driver is the
      only running driver from the task, and opA's driver is currently off thread
T4: opB memory reclaimer enter arbitration which find out its driver is not
       running so treat it as a memory arbitration issued from non-driver thread
       context so it doesn't put the driver in suspension state.
T5: when task tries to reclaim and hangs in task pause wait as running driver
       thread will never go down to zero.

This PR fix the issue by introducing a DriverThreadContext which contains the
driver ctx of a running driver. It is set to a thread local variable when a driver
starts execution on a driver thread and clears when the driver leaves the driver
thread. And provides to access the thread local variable to help detects if a
thread is running a driver and access the current driver data structure if needs.
First used in operator reclaimer enter arbitration handling to check if needs to
put the driver into suspension state. We simply needs to put the driver into
suspension state if it is set in thread local variable. Later on we might integrate
the per-op state such runtime stats into DriverThreadContext to consolidate
such driver/op execution contexts into one data structure to ease management.

Unit test is added to reproduce the deadlock and verified the fix.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6625

Reviewed By: mbasmanova

Differential Revision: D49408041

Pulled By: xiaoxmeng

fbshipit-source-id: 7a16c0425b99e637193097084409ac3ada334f80