Ensure RowVector::resize also resizes the children. (#6543)
Summary:
Currently when we call resize on a rowVector we do not resize its children. This creates problems for example in ContainerRowSerde where we resize the parent row (but the children still have the old size) and the new rows are not set to null. In general ensuring children are of same size as parent in resize helps to ensure vector is in a valid state.

There will be a subsequent PR that handles row vectors having null children, which isnt handled in this PR yet.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6543

Test Plan: Imported from GitHub, without a `Test Plan:` line.hee

Reviewed By: laithsakka, DanielMunozT

Differential Revision: D49241838

Pulled By: kgpai

fbshipit-source-id: fdbad576320124aa56202c5805163ded481d2bb8