Fix ArbitraryBuffer::empty() (#6644)
Summary:
`ArbitraryBuffer::empty()` returns true if there is no buffered page,
including two cases: no page exists, and only an end-marker exists.
For scenarios where the arbitrary buffer is marked as "no-more-data"
but there are still buffered pages in it, it should return false.

Because it is not a public API, this fix will not affect current users.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6644

Reviewed By: mbasmanova

Differential Revision: D49447208

Pulled By: xiaoxmeng

fbshipit-source-id: 5975015051b4c2705d3e30fb17e65e03fe29933d