Rename rows_ to ranges_ in detail::Destination (#6693)
Summary:
The `rows_` member of `detail::Destination` is misleading because it is
actually a list of `IndexRange`(an `IndexRange` represents a row-range).
In an `IndexRange`, there may be only one row, but it may also include
multiple consecutive rows(in the scenario of only one partition).

To avoid misunderstandings and improve readability, `rows_` is renamed
to `ranges_`, and `row_` is renamed to `rangeIdx_`.

No functional change in the PR.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6693

Reviewed By: kewang1024

Differential Revision: D49546875

Pulled By: xiaoxmeng

fbshipit-source-id: fa87b3c948835d9f4273e4c29fec38da580363b8