Fix handling of top level nulls in the input to Reduce function (#5939)
Summary:
Fixes https://github.com/facebookincubator/velox/issues/5768
Fix handling of top level nulls in the input to Reduce function. Currently, the nulls are being passed as rows to be operated on to the lambda functions in Reduce. This patch fixes this by ensuring only the non null Rows are passed to the lambda and nulls are added to the output.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5939

Test Plan: Adds a unit test for Reduce function.

Reviewed By: kevinwilfong

Differential Revision: D49084962

Pulled By: bikramSingh91

fbshipit-source-id: 581962521bbfe0c218e4b5bff6ca9b081a155505