Revert D49546239: Multisect successfully blamed "D49546239: Fix undefined behavior in bitwise_arithmetic_shift_right" for test or build failures (#6762)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6762

This diff is reverting D49546239
D49546239: Fix undefined behavior in bitwise_arithmetic_shift_right by laithsakka has been identified to be causing the following test or build failures:

Tests affected:
- [stylus/xstream/query/integration_test:e2e_test_builtin - test_bitwise (stylus.xstream.query.integration_test.e2e_test_builtin.XStreamBuiltinTest)](https://www.internalfb.com/intern/test/562950055334827/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3139628
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: laithsakka

Differential Revision: D49673433

fbshipit-source-id: 76d8305a0b584ce3609b2e667e7c0613cdf7e8d2