Support ordered memory pool reclaim (#6810)
Summary:
In Meta internal arbitration and spilling shadowing test, we found that without
enforcing reclaim order from a plan node, then the arbitrator always reclaim
memory follow the memory pool hierarchy, that is, always first reclaim from the
first child, second and so on. This might generates small files as the first operator
has less amount of in-memory state than the other which is not efficient as for
aggregation, we want to reduce the number of files as each file is a sorted run.

This PR optimizes this by sort the memory pool based on its reserved memory
before the actual memory reclaimation.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6810

Reviewed By: tanjialiang

Differential Revision: D49764544

Pulled By: xiaoxmeng

fbshipit-source-id: 298dea805b5e14dce3e06d218230efebe0d2cccf