Prefetch units should only contain relevant stripes (#6881)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6881

We don't want to return prefetch units for stripes that are out of range of the current reader.

A row reader can be given a byte range to read from, and the reader will only ever read from the stripes that are included in that byte range.

The reader will keep the `firstStripe` and (past) `lastStripe` variables showing its relevant stripes range.

We only want to return prefetch units in that range. So that the first prefetch unit points to `firstStripe` and the last is the stripe before `lastStripe`.

Reviewed By: Sullivan-Patrick

Differential Revision: D49889856

fbshipit-source-id: c8284f6ef61cdac284138abf707998e019d119c2