Fix null input handling in Reduce function (#6886)
Summary:
Reduce currently used the BaseVector's addNulls to add nulls at the end
to the localResult. This API is very basic and does not handle the variety
of cases that Expr::addNulls is designed to handle. For eg, if the target
vector is smaller than the input rows, or has a different encoding or is
empty. This change moves the Expr::addNulls() API to EvalCtx to make
it easily accessible to vector functions like Reduce without pulling in the
complete Expr header file and finally employing it to add nulls.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6886

Test Plan: Added unit tests

Reviewed By: mbasmanova

Differential Revision: D49897743

Pulled By: bikramSingh91

fbshipit-source-id: 22c0f0e636c5e1c53a1df32f9c5aabcf4277c949