Optimize addNulls for row vector. (#6856)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6856

One PR #6543 is re-landed, resizing row vectors will be expensive (children will be resized). #6543 was reverted due to other bugs that are exposed by it.
For the context of addNulls we can optimize that function by adding top level nulls in the row vector. instead of calling resize, this way we know #6543 will not add performance overhead for that path.
Moreover, this will mitigate the issues of resize for this location for row vectors. And will fix many fuzzer issues.

There are other optimizations that can be added to this function such as avoid calling ensnrewriable for complex types, by wrapping results in dictionary but those can be addressed in different diff.

Reviewed By: bikramSingh91, mbasmanova

Differential Revision: D49841643

fbshipit-source-id: 0f726001cf0fabb50deef7e90889cff06a3fac35