De-flake tests that verify task failures (#6907)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6907

Tests that run multi-threaded tasks that are expected to fail sometimes crash.

```
F1005 01:30:34.221653 2215122 DefaultKeepAliveExecutor.h:145] Check failed: keepAliveCount > 0
```

This happens because the Task and the Executor are destroyed while some drivers are still running. A fix is to wait for all drivers to complete before allowing the Task to go out of scope.

Reviewed By: xiaoxmeng

Differential Revision: D49947841

fbshipit-source-id: 10d4bb311e73aedb5e624ee885edf716601d30fc