[native pos] Report estimated size and row count of broadcast data
Report an exact number of rows broadcasted along with an estimate of broadcast
size in bytes. The estimate is based on serialized data size and roughly
matches the size of the broadcast file. This size may be different from the
size of the hash table that will be built from the broadcast data on the
reducer.

The estimate of the broadcast size will be used to check whether total size of
broadcast data is within a configured limit. If not, the map stage will be
retried with broadcast disabled.