Fix another tricky bug in mapFromEntries. (#6930)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6930

In the function mapFromEntries, if any row entry is null we should  set
 mutableSizes[row] = 0, to make sure the final vector is valid.

A problem is that errors that happens while iterator on those rows for a given
array at a given index, can prevent that check to happen for next rows.

For example: consider the array [(null,1), null], we should set sizes[0] = 0,
but we do not because the check not he null will throw and exit the loop.

This diff fix that issue and add a test case.

fix https://github.com/facebookincubator/velox/issues/6929

Reviewed By: mbasmanova

Differential Revision: D50000146

fbshipit-source-id: 7e5d3d058f37291390fc7ed84201ff42700403ee