Return FlatVector from DWRF string dictionary reader if selectivity is low (#6865)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6865

When the selectivity is low, we only need to read a very small portion
of the dictionary values, and copying the whole dictionary becomes more
expensive than just copying the selected `StringView`s.  We put in a condition
similar to Presto Java that if the selectivity of current batch is low enough,
we copy the selected result into a `FlatVector` directly.

Also add a small helper function `SelectiveColumnReader::resultNulls()` to
reduce repetition.

Reviewed By: oerling

Differential Revision: D49814314

fbshipit-source-id: 03ba3f8160af09c701eeecf963444290b98580c0