Remove storageFormat from dataDirectory in QueryRunners
The Java HiveQueryRunner takes a program argument as the input
dataDirectory so the users can persist any data in any format at this
location. However PrestoNativeQueryRunnerUtils always adds a
storageFormat string to the end of the input dataDirectory, and this
makes the existing user defined data directories unusable and previously
created schemas and data files invisible. The QueryRunners shall allow
the users to designate the data directory by themselves. This commit
removes the storageFormat from dataDirectory path.