Track result sizes of partial aggregate evaluation in HBO
Until now HBO tracked only the final input and output sizes of evaluating the aggregate, but not the result of partial agg evaluation.
This may lead to incorrect estimation of data size reduction when the final aggregate reduces data size, but the partial aggregate does not.
This changelist adds the ability to track input and output sizes of the partial aggregate to help make a better decision of when to split an aggregate
into partial and final.

This change adds a new field aggregationId to AggregationNode to track split partial/final aggregate pairs, and uses that during history tracking to record partial agg execution details.

The new statistic estimates now contains the following for Final aggregation nodes:

      "partialAggregationStatsEstimate" : {
        "inputBytes" : 0.0,
        "outputBytes" : 23976.0,
        "inputRowCount" : 5.6717574E7,
        "outputRowCount" : 2664.0
      }

This CL also changes the PushPartialAggregationThroughExchange optimizer to use the partial agg statistics when available.

In addition, the following modifications to the PushPartialAggregationThroughExchange optimization were made:

- when using partial aggregation statistics, apply optimization to multi-key aggregates. The original optimization only triggers for single-key aggregate nodes.
- Use rows instead of bytes when use_partial_aggregation_history flag is on