Silently drop tables with missing filesystem metadata
Adds two utility methods to bypass the previous "Table metadata is
missing" error thrown when attempting to drop a table for which a
metadata file or folder is missing from the underlying file system.
The utility methods eat the TableNotFoundException and continue
with the drop task.

Adds a test for this case as well.

Signed-off-by: kiersten-stokes <kierstenstokes@gmail.com>