Ensure AsciiInfo is updated during resize() & prepareForReuse(). (#6949)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6949

We are seeing rare crashes and not so rare ASAN crashes during
Vector copy() when trying to process SimpleVector's AsciiInfo.
It happens so during the Vector recycling we don't touch AsciiInfo at all
and the new vector might have AsciiInfo from the previous one.
When accessing AsciiInfo it might be smaller than the vector and we are
accessing memory outside of the buffer.

To fix this we do the following:
1. Call SimpleVector<StringView>::resizeIsAsciiIfNotEmpty() from FlatVector<T>::resize().
2. Call SimpleVector<StringView>::resizeIsAsciiIfNotEmpty() from ConstantVector::resize().
3. Check if AsciiInfo is at least size of the SimpleVector in SimpleVector::validate().
4. Ensure we use vector's length_ when allocating capacity for ascii bit vector, no matter what selectivity vector is passed.

Reviewed By: kgpai

Differential Revision: D50051589

fbshipit-source-id: c65dda6cc8a68b184637e7060e1e0ab92e273bfd