Fix null check in bloom_filter Spark aggregate (#7135)
Summary:
According to the DecodedVector::mayHaveNulls definition, it's uncertain whether there are nulls when `mayHaveNulls` is true.
https://github.com/facebookincubator/velox/blob/e8e84e38444c9292c2a79e0ea5c72a5c6312482f/velox/vector/DecodedVector.h#L184-L188

In Gluten, we found with current implementation, some of the Spark TPCDS queries that uses the BloomFilter aggregate fails because the `bloom_filter_agg` incorrectly fail based on the `mayHaveNulls` flag. We should remove the original check and, if `mayHaveNulls` is true, add a check for each row individually.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7135

Reviewed By: xiaoxmeng

Differential Revision: D50444224

Pulled By: mbasmanova

fbshipit-source-id: ccb4dce74136542f1b3608f59b25b8864c85fe7a