[native pos] Disable old task cleanup routine
PrestoServer contains oldTaskCleanup routine which cleans up tasks which are
not running or have not received heart-beat in a long while. This is meant to
detect zombie or stuck tasks, this is not relevant in Presto-On-Spark as
Spark speculative execution takes care of it.

In production we have seen legitimate tasks getting cleanup especially jobs
which are long running or have skew, disabling old task cleanup fixes the
problem.